requirements.txt
yfinance>=0.2.36
pandas>=2.2.0
pyyaml>=6.0.1
resend>=0.8.0
python-dotenv>=1.0.1
